requirements.txt
strawberry-graphql[debug-server]
flask